:+1: Isolate `function` module from `vim`
